paysages: Finally fixed camera transformation !
git-svn-id: https://subversion.assembla.com/svn/thunderk/paysages@249 b1fd45b6-86a6-48da-8261-f70d1f35bdcc

+++ gui_qt/widgetwanderer.cpp
@@ -189,7 +189,7 @@ void WidgetWanderer::resizeGL(int w, int h)
 {
     double ratio = (double)h / (double)w;
 
-    glViewport(0, 0, (GLint)w, (GLint)h);
+    glViewport(0, 0, w, h);
 
     glMatrixMode(GL_PROJECTION);
     glLoadIdentity();
